+++ rollup.config.js
@@ -2,7 +2,6 @@ import svelte from 'rollup-plugin-svelte';
 import resolve from 'rollup-plugin-node-resolve';
 import commonjs from 'rollup-plugin-commonjs';
 import { terser } from 'rollup-plugin-terser';
-import babel from 'rollup-plugin-babel';
 import sass from 'node-sass';
 
 const production = !process.env.ROLLUP_WATCH;
@@ -36,9 +35,6 @@ const shared = {
 		}),
 		resolve(),
 		commonjs(),
-		// babel({
-        //     exclude: 'node_modules/**'
-		// }),
 		production && terser()
 	]
 }