+++ src/App.svelte
@@ -94,6 +94,23 @@
 			labeltext="An example checkbox with some additional event handling of clicks inside">
 			<input slot="checkboxelement" type="checkbox"/>
 		</zoo-log-checkbox>
+		<zoo-log-radio valid="{inputState}" errormsg="errormsg" infotext="infotext">
+			<template>
+				<input type="radio" id="contactChoice1" name="contact" value="email" disabled>
+				<label for="contactChoice1">Email</label>
+				<input type="radio" id="contactChoice2" name="contact" value="phone">
+				<label for="contactChoice2">Phone</label>
+				<input type="radio" id="contactChoice3" name="contact" value="mail">
+				<label for="contactChoice3">Mail</label>
+			</template>
+		</zoo-log-radio>
+
+		<zoo-log-radio valid="{inputState}" errormsg="errormsg" infotext="infotext">
+			<input type="radio" id="contactChoice4" name="contact" value="email" disabled>
+			<label for="contactChoice4">Email</label>
+			<input type="radio" id="contactChoice5" name="contact" value="phone">
+			<label for="contactChoice5">Phone</label>
+		</zoo-log-radio>
 	</form>
 	<div class="buttons">
 		<zoo-log-button type="hot" size="medium" on:click="{changeState}">